Add cleanup-on-exit option to multus config

+++ pkg/server/config/manager.go
@@ -43,6 +43,7 @@ type Manager struct {
 	multusConfigDir            string
 	multusConfigFilePath       string
 	readinessIndicatorFilePath string
+	cleanupConfigOnExit        bool
 	primaryCNIConfigPath       string
 }
 
@@ -123,6 +124,7 @@ func newManager(config MultusConf, defaultCNIPluginName string) (*Manager, error
 		multusConfigFilePath:       filepath.Join(config.CniConfigDir, multusConfigFileName),
 		primaryCNIConfigPath:       filepath.Join(config.MultusAutoconfigDir, defaultCNIPluginName),
 		readinessIndicatorFilePath: config.ReadinessIndicatorFile,
+		cleanupConfigOnExit:        *config.CleanupConfigOnExit,
 	}
 
 	if err := configManager.loadPrimaryCNIConfigFromFile(); err != nil {
@@ -159,8 +161,10 @@ func (m *Manager) Start(ctx context.Context, wg *sync.WaitGroup) error {
 			_ = logging.Errorf("error watching file: %v", err)
 		}
 		logging.Verbosef("ConfigWatcher done")
-		logging.Verbosef("Delete old config @ %v", multusConfigFile)
-		os.Remove(multusConfigFile)
+		if m.cleanupConfigOnExit {
+			logging.Verbosef("Delete old config @ %v", multusConfigFile)
+			os.Remove(multusConfigFile)
+		}
 	}()
 
 	return nil
